distributed_test/requirements.txt
ray[default]
absl-py
jax[tpu]
-f https://storage.googleapis.com/jax-releases/libtpu_releases.html